Operator TakeTimed
